Removed old telepad impl and refactored telepad package.
