+++ src/main/java/crazypants/enderio/ClientProxy.java
@@ -56,9 +56,9 @@ import crazypants.enderio.teleport.TravelController;
 import crazypants.enderio.teleport.anchor.TileTravelAnchor;
 import crazypants.enderio.teleport.anchor.TravelEntitySpecialRenderer;
 import crazypants.enderio.teleport.telepad.ItemLocationPrintout;
-import crazypants.enderio.teleport.telepad.TelePadSpecialRenderer;
 import crazypants.enderio.teleport.telepad.TeleportEntityRenderHandler;
 import crazypants.enderio.teleport.telepad.TileTelePad;
+import crazypants.enderio.teleport.telepad.render.TelePadSpecialRenderer;
 import crazypants.util.ClientUtil;
 import net.minecraft.client.Minecraft;
 import net.minecraft.creativetab.CreativeTabs;
