+++ src/main/java/crazypants/enderio/teleport/telepad/BlockTelePad.java
@@ -16,6 +16,14 @@ import crazypants.enderio.teleport.ContainerTravelAccessable;
 import crazypants.enderio.teleport.ContainerTravelAuth;
 import crazypants.enderio.teleport.GuiTravelAuth;
 import crazypants.enderio.teleport.anchor.BlockTravelAnchor;
+import crazypants.enderio.teleport.telepad.gui.ContainerTelePad;
+import crazypants.enderio.teleport.telepad.gui.GuiAugmentedTravelAccessible;
+import crazypants.enderio.teleport.telepad.gui.GuiTelePad;
+import crazypants.enderio.teleport.telepad.packet.PacketOpenServerGui;
+import crazypants.enderio.teleport.telepad.packet.PacketTeleport;
+import crazypants.enderio.teleport.telepad.packet.PacketUpdateCoords;
+import crazypants.enderio.teleport.telepad.render.BlockType;
+import crazypants.enderio.teleport.telepad.render.TelePadRenderMapper;
 import net.minecraft.block.Block;
 import net.minecraft.block.properties.IProperty;
 import net.minecraft.block.properties.PropertyEnum;
