+++ src/main/java/crazypants/enderio/teleport/telepad/TileTelePad.java
@@ -23,6 +23,9 @@ import crazypants.enderio.machine.PacketPowerStorage;
 import crazypants.enderio.network.PacketHandler;
 import crazypants.enderio.teleport.TeleportUtil;
 import crazypants.enderio.teleport.anchor.TileTravelAnchor;
+import crazypants.enderio.teleport.telepad.packet.PacketTeleport;
+import crazypants.enderio.teleport.telepad.packet.PacketUpdateCoords;
+import crazypants.enderio.teleport.telepad.render.BlockType;
 import info.loenwind.autosave.annotations.Store;
 import net.minecraft.entity.Entity;
 import net.minecraft.entity.player.EntityPlayer;
