+++ src/main/java/crazypants/enderio/teleport/telepad/gui/ContainerTelePad.java
@@ -1,4 +1,4 @@
-package crazypants.enderio.teleport.telepad;
+package crazypants.enderio.teleport.telepad.gui;
 
 import java.awt.Point;
