+++ src/main/java/crazypants/enderio/teleport/telepad/packet/PacketUpdateCoords.java
@@ -1,9 +1,10 @@
-package crazypants.enderio.teleport.telepad;
+package crazypants.enderio.teleport.telepad.packet;
 
 import com.enderio.core.common.network.MessageTileEntity;
 import com.enderio.core.common.util.BlockCoord;
 
 import crazypants.enderio.EnderIO;
+import crazypants.enderio.teleport.telepad.ITileTelePad;
 import io.netty.buffer.ByteBuf;
 import net.minecraft.tileentity.TileEntity;
 import net.minecraftforge.fml.common.network.simpleimpl.IMessage;
