+++ src/main/java/crazypants/enderio/teleport/telepad/render/TelePadRenderMapper.java
@@ -1,4 +1,4 @@
-package crazypants.enderio.teleport.telepad.oldimpl;
+package crazypants.enderio.teleport.telepad.render;
 
 import java.util.Collections;
 import java.util.EnumMap;
@@ -6,19 +6,9 @@ import java.util.List;
 
 import javax.annotation.Nonnull;
 
-import net.minecraft.block.Block;
-import net.minecraft.block.state.IBlockState;
-import net.minecraft.item.ItemStack;
-import net.minecraft.tileentity.TileEntity;
-import net.minecraft.util.BlockRenderLayer;
-import net.minecraft.util.math.BlockPos;
-import net.minecraft.util.EnumFacing;
-import net.minecraft.world.IBlockAccess;
-import net.minecraftforge.fml.relauncher.Side;
-import net.minecraftforge.fml.relauncher.SideOnly;
-
 import org.apache.commons.lang3.tuple.Pair;
 
+import crazypants.enderio.EnderIO;
 import crazypants.enderio.render.EnumRenderMode;
 import crazypants.enderio.render.IBlockStateWrapper;
 import crazypants.enderio.render.ICacheKey;
@@ -26,45 +16,46 @@ import crazypants.enderio.render.IOMode.EnumIOMode;
 import crazypants.enderio.render.IRenderMapper;
 import crazypants.enderio.render.pipeline.ItemQuadCollector;
 import crazypants.enderio.render.pipeline.QuadCollector;
+import crazypants.enderio.teleport.telepad.BlockTelePad;
+import net.minecraft.block.Block;
+import net.minecraft.block.state.IBlockState;
+import net.minecraft.item.ItemStack;
+import net.minecraft.util.BlockRenderLayer;
+import net.minecraft.util.EnumFacing;
+import net.minecraft.util.math.BlockPos;
+import net.minecraft.world.IBlockAccess;
+import net.minecraftforge.fml.relauncher.Side;
+import net.minecraftforge.fml.relauncher.SideOnly;
 
-public class TelepadRenderMapperOld implements IRenderMapper.IBlockRenderMapper.IRenderLayerAware, IRenderMapper.IItemRenderMapper.IItemStateMapper {
+public class TelePadRenderMapper implements IRenderMapper.IBlockRenderMapper.IRenderLayerAware, IRenderMapper.IItemRenderMapper.IItemStateMapper {
 
-  public static final TelepadRenderMapperOld instance = new TelepadRenderMapperOld();
+  public static final TelePadRenderMapper instance = new TelePadRenderMapper();
 
   private static final EnumRenderMode GLASS_TOP_MODEL = EnumRenderMode.FRONT_ON;
   private static final EnumRenderMode FULL_MODEL = EnumRenderMode.FRONT_EAST;
   private static final EnumRenderMode SINGLE_MODEL = EnumRenderMode.FRONT;
   private static final EnumRenderMode SINGLE_MODEL_INVENTORY = EnumRenderMode.FRONT_SOUTH;
 
-  protected TelepadRenderMapperOld() {
+  protected TelePadRenderMapper() {
   }
 
   @Override
   @SideOnly(Side.CLIENT)
   public List<IBlockState> mapBlockRender(IBlockStateWrapper state, IBlockAccess world, BlockPos pos, BlockRenderLayer blockLayer,
                                           QuadCollector quadCollector) {
-    TileEntity tileEntity = state.getTileEntity();
 
-    if (tileEntity instanceof TileTelePadOld) {
-      TileTelePadOld telePad = (TileTelePadOld) tileEntity;
-
-      switch (blockLayer) {
-      case SOLID:
-        if (!telePad.inNetwork()) {
+    if (state.getBlock() == EnderIO.blockTelePad) {
+      if (blockLayer == BlockRenderLayer.SOLID) {
+        BlockType type = state.getValue(BlockTelePad.BLOCK_TYPE);
+        if (type == BlockType.SINGLE) {
           return Collections.singletonList(state.getState().withProperty(EnumRenderMode.RENDER, SINGLE_MODEL));
-        } else if (telePad.isMaster()) {
+        } else if (type == BlockType.MASTER) {
           return Collections.singletonList(state.getState().withProperty(EnumRenderMode.RENDER, FULL_MODEL));
         } else {
           return null;
         }
-      case TRANSLUCENT:
-        if (telePad.inNetwork() && telePad.isMaster()) {
-          return Collections.singletonList(state.getState().withProperty(EnumRenderMode.RENDER, GLASS_TOP_MODEL));
-        }
-      default:
-        return null;
       }
-
+      return null;
     } else if (blockLayer == BlockRenderLayer.CUTOUT) {
       return Collections.singletonList(state.getState().withProperty(EnumRenderMode.RENDER, SINGLE_MODEL));
     }
